+++ shared/src/commonMain/kotlin/com/lwtor/xhunter/data/ComicRepository.kt
@@ -1,0 +1,8 @@
+package com.lwtor.xhunter.data
+
+import com.lwtor.xhunter.ui.home.HomeComic
+import com.lwtor.xhunter.ui.home.HomeSubTab
+
+interface ComicRepository {
+    suspend fun getComics(tab: HomeSubTab): List<HomeComic>
+}